[MOV] im_livechat: move WebsiteLivechatMessage outside of -
-public_livechat.js

Task-2212347
